Feature/omnibus q1 (#10)
* Replace Chrome with Chromium

Reflects that these APIs ship in Chromium, not just Chrome, and that there are other contributors and vendors who may pull in from Chromium.

Resolves #8

* Clarify release timing

Added 'Stable' in from of 'X days ago' and 'in X days'

Resolves #6

* Update Prettier settings

Fixs the line length to 100 charcters, and updates relevant code to match

* Change Chrome to Chromium

Missed in the initial change

* Add deploy date to info

Resolves #2

* Add Explainers, Feature Flags, and Specs

* Clean up ship groupings

If ship date is earlier than OT end date, there’s likely a data problem w/the ship date, so remove shipping. Update other blocks accordingly. Also moves anything with a set shipping date of any sort to “started”

* WIP: Add TODOs for android/desktop differentation

A quick survey of the values here doesn’t show that they’re always more accurate than the information in the tickets so it’s unclear how to best incorporate these into the information display yet

* Add feedback links to info toggle

Resolves #4

* Add bug filtering

This was always meant to be a feature but for some reason it never got imlemented.

Resolves #7

* Fix double scrollbars

Resolves #3

* Improve Spec Display

Adds a copy button, shows the origin

* Change stripe color

The yellow didn’t look good with the new Chromium logo, so swapped it out for the chromium light blue